finish personal duplication formula

+++ app/models/personal_duplication_formula.rb
@@ -15,8 +15,13 @@ class PersonalDuplicationFormula < DuplicationFormula
   def delete_old_setting
     if self.group == "applied"
       @personal_old_settings = PersonalDuplicationFormula.find(:all, :conditions => ["duplication_formulas.group != ?", "default"])
-        @personal_old_settings.each do |i|
-        i.destroy
+      @personal_old_settings.each do |i|
+         if i != self
+           i.destroy
+           i.duplication_formula_details.each do |j|
+             j.destroy
+           end
+         end
       end
     end
